* Move `theme.js`
